refactor(fpss): extract events.rs
Move FpssData, FpssControl, FpssEvent, and the internal IoCommand enum from
fpss/mod.rs into a dedicated events submodule. Pure move -- no behavior
change. Public event types are re-exported from fpss so existing imports
(thetadatadx::fpss::FpssEvent etc.) remain stable.
